Add joker.markup namespace
Patterned on Hiccup, this namespace allows for rendering of XHTML (and variants) markup from vectors and strings.

It is not a direct port of Hiccup, as it is stricly interpretive, but should be quite sufficient for generating blog posts or XML configuration files, or other similar operations in Joker's intended domain.