Fix deps.joke to use source filename directly
Instead of guessing which argument, in the original invocation of Joker,
has the filename, or even using the Joker-provided *command-line-args*
variable, just use *file* directly (it's a new-ish feature).

This allows the test to pass even if Joker options (not options to the
Joker script being invoked) are provided. E.g. this works:

  ./joker --evaluate tests/run-eval-tests.joke