Add joker.core to *loaded-libs* explicitly
Because core.joke doesn't start out with `(ns joker.core ...)`, it doesn't naturally get added to `*loaded-libs*`.

`*loaded-libs*` used to list it explicitly, but for the sake of elegance and clarity sake, it seems better to leave it off the list, since (strictly speaking) joker.core is not actually loaded when Joker first starts up. (Although one could argue that `*loaded-libs*` itself does not exist in that case!)

Clarify the documentation for `*loaded-libs*` while at it.