Exclude generated code from builds of code-generating tools
This avoids a chicken-and-egg problem that comes up when e.g. changing whether a Joker type is a reference type.

It seems better to just avoid the problem, by not including code being generated via `go generate ./...` when building the tools doing the generation, than to document how to work around it.