Rename/refactor most of the Ensure/Assert functions
Currently the caller of Assert* is given a binary choice: pass an empty message, letting the API (if necessary) compose a useful one with the expected and actual type information; or, provide a complete message, requiring the caller to go to the expense of doing that even if (as is most often the case) it's not needed due to the type being in conformance.

Herein, the message is converted to a pattern (the default being "%s"), and the pattern is expanded only when needed, providing the current default message as the only argument to the pattern. (For most callers, this is sufficient; a few must still precompute a pattern based on runtime info, so if that is deemed an issue, arbitrary additional arguments could be supported via e.g. "...interface{}" and such.)

The result is simplified, as well as slightly faster, code. The resulting messages are different (e.g. compare the output of '(double "hey")'), but that -- surprisingly -- doesn't seem to affect the current set of tests.

Also, I've consistently been confused by the distinction between "Assert" and "Ensure". While I don't have a strong feeling about which term is used, it seems like only one should be (and "Assert" implies, to me, a mere check, not also the return of a value of the desired datatype).

So I've renamed "Assert<T>" to "EnsureObjectIs<T>", after renaming "Ensure<T>" to "EnsureArgIs<T>" (where appropriate). I think this'll make it a little easier to read code without resorting to looking at types_assert_gen.go so frequently.