Try another approach to reduce executable, not just code, size
The executables on amd64-linux and amd64-darwin got about 100K bytes
larger, despite the generated code having fewer LOC.

Guessing that this might be due primarily to a combination of varargs
and use of interface{} (the "...interface{}" args), this approach
removes the responsibility of panic()'ing from the lowest-level
type-ensurer function to the caller by adding an extra "ok" return
argument for the caller to test.

Since that requires returning both the desired object and the "ok"
bool, a nil value in the "not-ok" (ok==false) case must be returned,
and this is either "nil" or the result of constructing an empty object
of the desired type.

So this approach adds an "=" prefix to typenames passed, by object.go,
to gen_types.go, to indicate that they need an empty constructor vs
simply "nil" for their nil type.

The resulting executable is still 10K bytes larger on amd64-linux, but
that's only 10% of the previous delta.