Implement Java's Character.isWhitespace() More Accurately
This might nail it, but I'm not 100% sure. It brings Joker a lot closer to what Clojure allows (and disallows) in symbol/keyword names and in what is considered whitespace generally.

Also, this code strives to handle the common cases quickly via explicit short-circuiting. It's possible Go would figure that stuff out anyway, but why rely on it when the code remains clear enough?