PR feedback: Move new precision functions into joker.math
This seemed simple, but ran up against the fact that BigInt's `b` member is not exported.

Rather than kludge around that, it seemed better to be more consistent about having `Make<Type>` functions, where `<Type>` is `BigInt` and `BigFloat`, take the underlying (wrapped) types as arguments, like other `Make<Type>` functions, rather than making assumptions about whether those values are signed, unsigned, fit in a `int`, etc.

This, and the introduction of a `Precision` abstract type in `numbers.go` that `math_native.go` simply calls to determine the precision of an arbitrary `Number`, seems like a better division of responsibilities and knowledge among the relevant components.