Support explicit environment in joker.os/exec
This adds support for a `:env` hash key when calling `joker.os/exec`. When its value is `true`, the current environment (as might be returned by `joker.os/env` as a hash) is used for the new process's environment. When unspecified or its value is `false`, the behavior of Joker itself is unchanged, but the Go runtime might, as of 1.19, set the `PWD` environment variable, in that new environment, to the value of `:dir`, if specified, changing the behavior of some existing code.